# apps/openapi/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "My Catatan Kuliah",
    "version": "1.0.0",
    "description": "My Catatan Kuliah API"
  },
  "servers": [
    {
      "url": "/",
      "description": "Default server",
      "variables": {}
    }
  ],
  "paths": {
...
    "/api/notes": {
      "post": {
        "operationId": "notesServiceCreateNote",
        "description": "Create a new note with the specified details.",
        "parameters": [],
        "requestBody": {
          "description": "The details of the note to create",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotesCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The request has succeeded and a new resource has been created as a result.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotesCreateResponse"
                }
              }
            }
          },
          "400": {
            "description": "The server could not understand the request due to invalid syntax.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonErrorError400"
                }
              }
            }
          },
          "401": {
            "description": "Access is unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonErrorError401"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonErrorError500"
                }
              }
            }
          }
        },
        "tags": [
          "Notes"
        ]
      },
      "get": {
        "operationId": "notesServiceListNotes",
        "description": "List all notes accessible to the authenticated user.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotesListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Access is unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonErrorError401"
                }
              }
            }
          },
          "404": {
            "description": "The server cannot find the requested resource.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonErrorError404"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonErrorError500"
                }
              }
            }
          }
        },
        "tags": [
          "Notes"
        ]
      }
    },
    "/api/notes/{noteId}": {
      "get": {
        "operationId": "notesServiceGetNote",
        "description": "Get details of a specific note by its ID.",
        "parameters": [
          {
            "name": "noteId",
            "in": "path",
            "description": "The unique identifier of the note",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-oapi-codegen-extra-tags": {
              "validate": "required"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotesDetailResponse"
                }
              }
            }
          },
          "401": {
            "description": "Access is unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonErrorError401"
                }
              }
            }
          },
          "404": {
            "description": "The server cannot find the requested resource.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonErrorError404"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonErrorError500"
                }
              }
            }
          }
        },
        "tags": [
          "Notes"
        ]
      },
      "delete": {
        "operationId": "notesServiceDeleteNote",
        "description": "Delete a specific note by its ID.",
        "parameters": [
          {
            "name": "noteId",
            "in": "path",
            "description": "The unique identifier of the note",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-oapi-codegen-extra-tags": {
              "validate": "required"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "There is no content to send for this request, but the headers may be useful. "
          },
          "401": {
            "description": "Access is unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonErrorError401"
                }
              }
            }
          },
          "404": {
            "description": "The server cannot find the requested resource.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonErrorError404"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonErrorError500"
                }
              }
            }
          }
        },
        "tags": [
          "Notes"
        ]
      },
      "patch": {
        "operationId": "notesServiceUpdateNote",
        "description": "Update details of a specific note by its ID.",
        "parameters": [
          {
            "name": "noteId",
            "in": "path",
            "description": "The unique identifier of the note",
            "required": true,
            "schema": {
              "type": "string"
            },
            "x-oapi-codegen-extra-tags": {
              "validate": "required"
            }
          }
        ],
        "requestBody": {
          "description": "The details to update for the note",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NotesUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The request has succeeded.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotesUpdateResponse"
                }
              }
            }
          },
          "400": {
            "description": "The server could not understand the request due to invalid syntax.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonErrorError400"
                }
              }
            }
          },
          "401": {
            "description": "Access is unauthorized.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonErrorError401"
                }
              }
            }
          },
          "404": {
            "description": "The server cannot find the requested resource.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonErrorError404"
                }
              }
            }
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CommonErrorError500"
                }
              }
            }
          }
        },
        "tags": [
          "Notes"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CommonErrorError": {
        "description": "An error response with details about the failure.",
        "type": "object",
        "properties": {
          "id": {
            "description": "A unique identifier for this error instance",
            "type": "string",
            "example": "123e4567-e89b-12d3-a456-426614174000",
            "x-oapi-codegen-extra-tags": {
              "validate": "required"
            }
          },
          "reason": {
            "description": "A human-readable explanation of the error",
            "type": "string",
            "example": "Resource Not Found",
            "x-oapi-codegen-extra-tags": {
              "validate": "required"
            }
          },
          "status": {
            "description": "The HTTP status code associated with this error",
            "type": "string",
            "example": "Not Found",
            "x-oapi-codegen-extra-tags": {
              "validate": "required"
            }
          },
          "message": {
            "description": "An optional message providing more details about the error",
            "type": "string",
            "example": "The requested resource could not be found",
            "x-oapi-codegen-extra-tags": {
              "validate": "omitempty"
            }
          },
          "code": {
            "description": "An optional machine-readable error code for programmatic handling",
            "type": "integer",
            "format": "int32",
            "example": 404,
            "x-oapi-codegen-extra-tags": {
              "validate": "required,int32"
            }
          },
          "details": {
            "description": "Optional additional details about the error, such as missing resources or validation errors",
            "type": "object",
            "example": {
              "missingResource": "assignment with ID 123 not found"
            },
            "additionalProperties": {
              "type": "string"
            },
            "x-oapi-codegen-extra-tags": {
              "validate": "omitempty"
            }
          }
        },
        "required": [
          "id",
          "reason",
          "status",
          "code"
        ]
      },
      "CommonErrorError400": {
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/CommonErrorError"
          }
        ]
      },
      "CommonErrorError401": {
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/CommonErrorError"
          }
        ]
      },
      "CommonErrorError404": {
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/CommonErrorError"
          }
        ]
      },
      "CommonErrorError500": {
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/CommonErrorError"
          }
        ]
      },
...
      "NotesCreateRequest": {
        "type": "object",
        "properties": {
          "course_id": {
            "type": "string",
            "x-oapi-codegen-extra-tags": {
              "validate": "required"
            }
          },
          "title": {
            "type": "string",
            "x-oapi-codegen-extra-tags": {
              "validate": "required"
            }
          },
          "content": {
            "type": "string",
            "x-oapi-codegen-extra-tags": {
              "validate": "omitempty"
            }
          }
        },
        "required": [
          "course_id",
          "title"
        ]
      },
      "NotesCreateResponse": {
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/NotesDetailResponse"
          }
        ]
      },
      "NotesDetailResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "readOnly": true,
            "x-oapi-codegen-extra-tags": {
              "validate": "required"
            }
          },
          "course_id": {
            "type": "string",
            "x-oapi-codegen-extra-tags": {
              "validate": "required"
            }
          },
          "title": {
            "type": "string",
            "x-oapi-codegen-extra-tags": {
              "validate": "required"
            }
          },
          "created_by": {
            "type": "string",
            "readOnly": true,
            "x-oapi-codegen-extra-tags": {
              "validate": "required"
            }
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "readOnly": true,
            "x-oapi-codegen-extra-tags": {
              "validate": "required,datetime=2006-01-02T15:04:05Z07:00"
            }
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "readOnly": true,
            "x-oapi-codegen-extra-tags": {
              "validate": "required,datetime=2006-01-02T15:04:05Z07:00"
            }
          },
          "content": {
            "type": "string",
            "x-oapi-codegen-extra-tags": {
              "validate": "omitempty"
            }
          }
        },
        "required": [
          "id",
          "course_id",
          "title",
          "created_by",
          "created_at",
          "updated_at"
        ]
      },
      "NotesListResponse": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/NotesResponse"
        },
        "x-oapi-codegen-extra-tags": {
          "validate": "omitempty"
        }
      },
      "NotesResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "readOnly": true,
            "x-oapi-codegen-extra-tags": {
              "validate": "required"
            }
          },
          "course_id": {
            "type": "string",
            "x-oapi-codegen-extra-tags": {
              "validate": "required"
            }
          },
          "title": {
            "type": "string",
            "x-oapi-codegen-extra-tags": {
              "validate": "required"
            }
          },
          "created_by": {
            "type": "string",
            "readOnly": true,
            "x-oapi-codegen-extra-tags": {
              "validate": "required"
            }
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "readOnly": true,
            "x-oapi-codegen-extra-tags": {
              "validate": "required,datetime=2006-01-02T15:04:05Z07:00"
            }
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "readOnly": true,
            "x-oapi-codegen-extra-tags": {
              "validate": "required,datetime=2006-01-02T15:04:05Z07:00"
            }
          }
        },
        "required": [
          "id",
          "course_id",
          "title",
          "created_by",
          "created_at",
          "updated_at"
        ]
      },
      "NotesUpdateRequest": {
        "description": "The template for adding optional properties.",
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "x-oapi-codegen-extra-tags": {
              "validate": "omitempty"
            }
          },
          "content": {
            "type": "string",
            "x-oapi-codegen-extra-tags": {
              "validate": "omitempty"
            }
          }
        }
      },
      "NotesUpdateResponse": {
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/NotesDetailResponse"
          }
        ]
      },
...
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session"
      }
    }
  },
...
  "security": [
    {
      "ApiKeyAuth": []
    }
  ]
}